Database was created according to the specifications with the following structure: Entity name: Population Attributes: OIB VARCHAR(11) Name TEXT, Surname TEXT, Occupation TEXT, Date_Of_Birth TEXT, Height INT, Weight INT, Blood_Type VARCHAR(3) Additional_Information TEXT
The Date_Of_Birth attribute uses the TEXT data type which follows the YYY-MM-DD ISO8601 standard instead of INTEGER which would use Unix time. This makes date comparison more difficult, but representation in the database is .

Alternatively it can be changed to be INTEGER however when the data is being collected, the date() functions of SQLite3 would have t be used in order to transform the dates into readable formats. For more information, reference the documentation:
https://www.sqlite.org/lang_datefunc.html