Created packages got database, model and routes. In the database package, there is the database.go program which includes the funtions needed to connect to and close a databse connection. NOTE: The location of the database should not be hardcoded and changed with an environment variable or other alternative mean
The route program defines a function for setting up all the possible routes that the backend will have. Due to the controllers not being implemented, none have yet to be added

The controller will define what to do with specific requests. When requests arrive on the backend, the appropriate controller functions will be called depending on the route used.

The model includes population.go which mirrors the databases design without using an ORM like GORM (still debating whether or not to switch to it for ease of use)